Review comments
Also updated to rxnetty-0.5.2-rc.3
